backend/cloud_functions/gmail_enricher/requirements.txt
google-cloud-bigquery>=3.12.0
google-cloud-language>=2.11.0
google-auth>=2.23.0
google-cloud-aiplatform>=1.38.0
functions-framework>=3.4.0
